requirements.txt
langchain
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
chromadb
langchain-chroma
streamlit
beautifulsoup4
playwright
unstructured
langchain-groq
python-dotenv
rank-bm25